Hitboxes, Classes, Recoil
Hitboxes are now classified as either 'uni' or 'bi' (directional).
Unidirectional hitboxes cause the target to recoil in the direction that
the producer is facing in. Bidirectional hitboxes cause the target to
recoil away from the producer. 'Enemy' and 'Ally' are now subclasses of
'Fighter', as Fighters all share certain code. The player can now recoil
backwards and forwards.